feat: successful waiver upload to s3
