--- src/app/api/events/[eventID]/waiver/route.ts
@@ -1,217 +1,0 @@
-import { NextRequest, NextResponse } from "next/server";
-import { auth } from "@clerk/nextjs/server";
-import connectDB from "@/database/db";
-import User from "@/database/userSchema";
-import Event from "@/database/eventSchema";
-import Waiver from "@/database/waiverSchema";
-import mongoose from "mongoose";
-import { PDFDocument } from "pdf-lib";
-import { PdfReader } from "pdfreader";
-import { s3 } from "@/lib/s3";
-import { GetObjectCommand, PutObjectCommand } from "@aws-sdk/client-s3";
-
-export async function POST(req: NextRequest, { params }: { params: { eventID: string } }) {
-  await connectDB();
-  const { eventID } = params;
-
-  if (!mongoose.Types.ObjectId.isValid(eventID)) {
-    return NextResponse.json({ error: "Invalid event ID" }, { status: 400 });
-  }
-
-  const { userId } = await auth();
-  if (!userId) {
-    return NextResponse.json({ error: "Unauthorized. Please log in." }, { status: 401 });
-  }
-
-  const user = await User.findOne({ clerkID: userId });
-  if (!user) {
-    return NextResponse.json({ error: "User not found" }, { status: 404 });
-  }
-
-  // Get waiverID from the request body
-  const { signatureBase64, waiverID, participants } = await req.json(); // Accept waiverID and signatureBase64 as part of the request body
-  if (!signatureBase64 || !waiverID || !participants) {
-    return NextResponse.json({ error: "No signature, waiverID, or participants provided." }, { status: 400 });
-  }
-
-  try {
-    // Fetch the waiver document using the waiverID to get the fileKey
-    const waiver = await Waiver.findById(waiverID);
-    if (!waiver) {
-      return NextResponse.json({ error: "Waiver not found" }, { status: 404 });
-    }
-
-    const { fileKey } = waiver; // Get the fileKey from the waiver document
-
-    // Fetch the unsigned waiver from S3 using the fileKey
-    const getObjectParams = {
-      Bucket: process.env.AWS_BUCKET_NAME,
-      Key: fileKey, // The fileKey from the waiver document
-    };
-
-    const s3Object = await s3.send(new GetObjectCommand(getObjectParams));
-
-    // Read the file from S3 and convert it into a PDF byte array
-    const pdfBytes = await streamToBuffer(s3Object.Body); // Convert S3 stream to buffer
-
-    // Use pdfreader to extract text and find the word 'SIGNATURE' first, then fall back to 'sign'
-    const positions = await extractTextAndFindSign(pdfBytes);
-
-    // Choose the first position to place the signature
-    const signPosition = positions[0];
-
-    const pdfDoc = await PDFDocument.load(pdfBytes);
-    const page = pdfDoc.getPages()[signPosition.page - 1];
-    const pdfWidth = page.getWidth();
-    const pdfHeight = page.getHeight();
-
-    const pngImage = await pdfDoc.embedPng(signatureBase64);
-    const { width, height } = pngImage.scale(0.35); // Scale signature image
-
-    // Scale reader's x/y into PDF's coordinate space
-    const scaledX = (signPosition.x / 100) * 2 * pdfWidth;
-    const scaledY = pdfHeight - (signPosition.y / 100) * 2 * pdfHeight; // Invert y-axis
-
-    page.drawImage(pngImage, {
-      x: scaledX,
-      y: scaledY,
-      width,
-      height,
-    });
-
-    const modifiedPdf = await pdfDoc.save();
-
-    // Generate a unique file key for the signed version in S3
-    const signedFilename = `waiver-signed-${eventID}-${user._id}.pdf`;
-    const s3Key = `waivers/completed/${eventID}/${user._id}/${signedFilename}`;
-
-    // Upload the signed waiver to S3
-    const uploadParams = {
-      Bucket: process.env.AWS_BUCKET_NAME,
-      Key: s3Key,
-      Body: modifiedPdf,
-      ContentType: "application/pdf",
-    };
-
-    await s3.send(new PutObjectCommand(uploadParams));
-
-    // 1) Get presigned URL → 2) PUT to S3 → 3) return final URL + key + name
-    const uploadPDF = async (file: File, eventId: string): Promise<PDFInfo> => {
-      const presigned = await fetch(
-        `/api/s3/presigned-waiver?fileName=${encodeURIComponent(signedFilename)}&mimetype=application/pdf&type=completed&eventId=${eventId}`,
-      );
-      if (!presigned.ok) throw new Error("Presign failed");
-      const { uploadUrl, fileUrl, key } = await presigned.json();
-      await fetch(uploadUrl, {
-        method: "PUT",
-        body: file,
-        headers: { "Content-Type": "application/pdf" },
-      });
-      return { fileUrl, fileKey: key, fileName: file.name };
-    };
-
-    // Generate a publicly accessible URL to the uploaded file
-    const fileUrl = `https://${process.env.AWS_BUCKET_NAME}.s3.${process.env.AWS_REGION}.amazonaws.com/${s3Key}`;
-
-    // Loop through all participants and create individual waiver records
-    const waiverDocs = [];
-
-    for (const participant of participants) {
-      const isSelf = participant.userID === user.clerkID;
-
-      let childSubdocId = undefined;
-
-      if (!isSelf) {
-        // make sure child is valid/user's child
-        const child = user.children.find((child) => child._id.toString() === participant.userID);
-
-        if (!child) {
-          console.warn(`No matching child found for participant: ${participant.firstName} ${participant.lastName}`);
-          continue; // skip if child not found
-        }
-
-        childSubdocId = child._id;
-      }
-
-      // create a new waiver document
-      const participantWaiver = await Waiver.create({
-        fileKey: s3Key,
-        fileName: signedFilename,
-        uploadedBy: user._id,
-        belongsToUser: user._id,
-        childSubdocId: childSubdocId,
-        isForChild: !isSelf,
-        type: "completed",
-        templateRef: waiver._id,
-        eventId: new mongoose.Types.ObjectId(eventID),
-      });
-
-      // update the user scheme
-      if (isSelf) {
-        user.waiversSigned.push(participantWaiver._id);
-      } else if (childSubdocId) {
-        const child = user.children.id(childSubdocId);
-        if (child) {
-          child.waiversSigned.push(participantWaiver._id);
-        }
-      }
-
-      waiverDocs.push(participantWaiver);
-    }
-
-    await user.save();
-
-    return NextResponse.json({
-      message: "Waiver signed and uploaded!",
-      signedPdfUrl: fileUrl, // Return the S3 URL where the file can be accessed
-    });
-  } catch (error) {
-    console.error("Error processing waiver:", error);
-    return NextResponse.json({ error: "Failed to sign waiver." }, { status: 500 });
-  }
-}
-
-// Function to extract text and find positions for exact 'SIGNATURE' or fallback to 'sign'
-async function extractTextAndFindSign(pdfBytes: Buffer) {
-  return new Promise((resolve, reject) => {
-    const exactMatches: any[] = [];
-    const fallbackMatches: any[] = [];
-    let currentPage = 0;
-
-    new PdfReader().parseBuffer(pdfBytes, (err, item) => {
-      if (err) return reject(err);
-
-      if (item?.page) {
-        currentPage = item.page;
-      }
-
-      if (item && item.text) {
-        const text = item.text.trim();
-        const normalized = text.replace(/\s+/g, "").toLowerCase(); // Remove spaces for matching
-
-        if (/^signature$/.test(normalized)) {
-          console.log(`[MATCH: EXACT] Page ${currentPage}: "${text}"`);
-          exactMatches.push({ x: item.x, y: item.y, page: currentPage, text });
-        } else if (/^signa/.test(normalized)) {
-          console.log(`[MATCH: FALLBACK] Page ${currentPage}: "${text}"`);
-          fallbackMatches.push({ x: item.x, y: item.y, page: currentPage, text });
-        }
-      }
-
-      if (!item) {
-        const finalMatches = exactMatches.length > 0 ? exactMatches : fallbackMatches;
-        resolve(finalMatches);
-      }
-    });
-  });
-}
-
-// Helper function to convert S3 stream to buffer
-async function streamToBuffer(stream: any): Promise<Buffer> {
-  return new Promise((resolve, reject) => {
-    const chunks: any[] = [];
-    stream.on("data", (chunk) => chunks.push(chunk));
-    stream.on("end", () => resolve(Buffer.concat(chunks)));
-    stream.on("error", reject);
-  });
-}
